Finally have ride arrow lines behaving the way I want.
Capped the maximum number of lines to be drawn at 500 for performance reasons. May be worth adding an option to draw all lines to the options window?

Repurposed the "Show Top 100 Routes" such that it now caps the number of lines drawn when a station is selected to the 100 routes with the most riders.